+++ packages/multiple-choice/configure/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice-configure@0.8.5...@pie-element/multiple-choice-configure@0.8.6) (2019-04-08)
+
+**Note:** Version bump only for package @pie-element/multiple-choice-configure
+
+
+
+
+
+## [0.8.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice-configure@0.8.4...@pie-element/multiple-choice-configure@0.8.5) (2019-04-08)
+
+
+### Bug Fixes
+
+* Update state.model if there are any changes. ([610a9db](https://github.com/pie-framework/pie-elements/commit/610a9db))
+
+
+
+
+
 ## [0.8.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice-configure@0.8.3...@pie-element/multiple-choice-configure@0.8.4) (2019-04-04)