+++ packages/select-text/controller/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-controller@2.4.0...@pie-element/select-text-controller@2.4.1) (2019-04-09)
+
+**Note:** Version bump only for package @pie-element/select-text-controller
+
+
+
+
+
+# [2.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-controller@2.3.4...@pie-element/select-text-controller@2.4.0) (2019-04-09)
+
+
+### Features
+
+* partial scoring - use @pie-lib/controller-utils ([dac819f](https://github.com/pie-framework/pie-elements/commit/dac819f))
+
+
+
+
+
 ## [2.3.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/select-text-controller@2.3.3...@pie-element/select-text-controller@2.3.4) (2019-03-18)
 
 **Note:** Version bump only for package @pie-element/select-text-controller